chore: apply clippy fixes

+++ backend/src/scanners/similar_videos.rs
@@ -7,11 +7,11 @@ use crate::models::{FileGroup, ScanRequest, ScanResults, ScannedFile};
 use crate::scanners::{configure_common_data, make_stop_flag};
 
 pub fn run(request: ScanRequest) -> Result<ScanResults, String> {
-    let tolerance = request.tolerance.unwrap_or(5).clamp(0, 20) as i32;
+    let tolerance = request.tolerance.unwrap_or(5).clamp(0, 20);
     let duration = request
         .vid_hash_duration
         .unwrap_or(10)
-        .clamp(2, 60) as u32;
+        .clamp(2, 60);
     let crop_detect = request
         .crop_detect
         .as_deref()
